BOLTZMAN/streamui/requirements.txt
streamlit==1.28.2
streamlit-echarts==0.4.0
pika==1.2.0
minio==7.2.0
pandas==2.0.3
matplotlib==3.7.4
plotly==5.18.0
xlrd==2.0.1
redis==5.0.1
async_timeout==4.0.3
scikit-learn==1.3.2
mlxtend==0.23.0
xlrd==2.0.1
seaborn==0.13.0
pca==2.0.5
streamlit-keyup==0.2.2
python3-logstash==0.4.80
pymongo==4.6.1
elasticsearch==8.12.0